Remove unused react.svg image
Deleted a react.svg that was not being used in the project. This contributes to keeping the project lean and free of unused resources.

Added custom styles (Bootstrap) for form elements

This includes styles for general form controls and introduces floating labels and range input. The styles are structured as separate scss files and imported into bootstrap-utilities for easier manageability and better organization.

The updated styles and layouts will improve the look and feel of form controls across the application.